c.lib 1.174 (27.8.2004)
- The parameters of atan2() were swapped. Fixed.

- Merged fdlibm 5.3 changes with __kernel_tan(), __exp() and __pow()
  functions.


git-svn-id: file:///Users/olsen/Code/migration-svn-zu-git/logical-line-staging/clib2/trunk@14718 87f5fb63-7c3d-0410-a384-fd976d0f7a62

+++ library/math_exp.c
@@ -1,5 +1,5 @@
 /*
- * $Id: math_exp.c,v 1.1.1.1 2004-07-26 16:30:41 obarthel Exp $
+ * $Id: math_exp.c,v 1.2 2004-08-27 11:40:49 obarthel Exp $
  *
  * :ts=4
  *
@@ -194,7 +194,7 @@ __exp(double x)
 	    } 
 		else 
 		{
-			k  = invln2*x+halF[xsb];
+			k  = (int)(invln2*x+halF[xsb]);
 			t  = k;
 			hi = x - t*ln2HI[0];	/* t*ln2HI is exact here */
 			lo = t*ln2LO[0];
